+++ packages/expect/src/__tests__/utils.test.js
@@ -21,6 +21,7 @@ const {
 describe('getPath()', () => {
   test('property exists', () => {
     expect(getPath({a: {b: {c: 5}}}, 'a.b.c')).toEqual({
+      endPropIsDefined: true,
       hasEndProp: true,
       lastTraversedObject: {c: 5},
       traversedPath: ['a', 'b', 'c'],
@@ -28,6 +29,7 @@ describe('getPath()', () => {
     });
 
     expect(getPath({a: {b: {c: {d: 1}}}}, 'a.b.c.d')).toEqual({
+      endPropIsDefined: true,
       hasEndProp: true,
       lastTraversedObject: {d: 1},
       traversedPath: ['a', 'b', 'c', 'd'],
@@ -37,6 +39,7 @@ describe('getPath()', () => {
 
   test('property doesnt exist', () => {
     expect(getPath({a: {b: {}}}, 'a.b.c')).toEqual({
+      endPropIsDefined: false,
       hasEndProp: false,
       lastTraversedObject: {},
       traversedPath: ['a', 'b'],
@@ -46,6 +49,7 @@ describe('getPath()', () => {
 
   test('property exist but undefined', () => {
     expect(getPath({a: {b: {c: undefined}}}, 'a.b.c')).toEqual({
+      endPropIsDefined: true,
       hasEndProp: true,
       lastTraversedObject: {c: undefined},
       traversedPath: ['a', 'b', 'c'],
@@ -64,12 +68,14 @@ describe('getPath()', () => {
     }
 
     expect(getPath(new A(), 'a')).toEqual({
+      endPropIsDefined: true,
       hasEndProp: true,
       lastTraversedObject: new A(),
       traversedPath: ['a'],
       value: 'a',
     });
     expect(getPath(new A(), 'b.c')).toEqual({
+      endPropIsDefined: true,
       hasEndProp: true,
       lastTraversedObject: {c: 'c'},
       traversedPath: ['b', 'c'],
@@ -82,6 +88,7 @@ describe('getPath()', () => {
     A.prototype.a = 'a';
 
     expect(getPath(new A(), 'a')).toEqual({
+      endPropIsDefined: true,
       hasEndProp: true,
       lastTraversedObject: new A(),
       traversedPath: ['a'],
@@ -100,6 +107,7 @@ describe('getPath()', () => {
 
   test('empty object at the end', () => {
     expect(getPath({a: {b: {c: {}}}}, 'a.b.c.d')).toEqual({
+      endPropIsDefined: false,
       hasEndProp: false,
       lastTraversedObject: {},
       traversedPath: ['a', 'b', 'c'],