+++ packages/expect/src/matchers.ts
@@ -648,11 +648,12 @@ const matchers: MatchersObject = {
     }
 
     const result = getPath(object, keyPath);
-    const {lastTraversedObject, hasEndProp} = result;
+    const {lastTraversedObject, endPropIsDefined, hasEndProp} = result;
 
-    const pass = valuePassed
-      ? equals(result.value, value, [iterableEquality])
-      : hasEndProp;
+    const pass =
+      valuePassed && endPropIsDefined
+        ? equals(result.value, value, [iterableEquality])
+        : hasEndProp;
 
     const traversedPath = result.traversedPath.join('.');